add paint & attack system
탄알 시스템 , 착탄 후 페인트 , collision 후 공격 들어가야함

+++ Assets/Scripts/Character/Config/HealthConfigSO.cs
@@ -3,8 +3,8 @@
 [CreateAssetMenu(fileName = "HealthConfig", menuName = "EntityConfig/Health Config")]
 public class HealthConfigSO : ScriptableObject
 {
-	[SerializeField] private int _initialHealth;
+	[SerializeField] private float _initialHealth;
 
-	public int InitialHealth => _initialHealth;
+	public float InitialHealth => _initialHealth;
 
 }
