+++ Assets/Scripts/Character/HealthSO.cs
@@ -8,28 +8,28 @@
 public class HealthSO : ScriptableObject
 {
 	[Tooltip("The initial health")]
-	[SerializeField][ReadOnly] private int _maxHealth;
-	[SerializeField][ReadOnly] private int _currentHealth;
+	[SerializeField][ReadOnly] private float _maxHealth;
+	[SerializeField][ReadOnly] private float _currentHealth;
 
-	public int MaxHealth => _maxHealth;
-	public int CurrentHealth => _currentHealth;
+	public float MaxHealth => _maxHealth;
+	public float CurrentHealth => _currentHealth;
 
-	public void SetMaxHealth(int newValue)
+	public void SetMaxHealth(float newValue)
 	{
 		_maxHealth = newValue;
 	}
 
-	public void SetCurrentHealth(int newValue)
+	public void SetCurrentHealth(float newValue)
 	{
 		_currentHealth = newValue;
 	}
 	
-	public void InflictDamage(int DamageValue)
+	public void InflictDamage(float DamageValue)
 	{
 		_currentHealth -= DamageValue;
 	}
 
-	public void RestoreHealth(int HealthValue)
+	public void RestoreHealth(float HealthValue)
 	{
 		_currentHealth += HealthValue;
 		if(_currentHealth > _maxHealth)
